requirements.txt
Click==7.0
Flask==1.0.2
Flask-SocketIO==3.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
opencv-python==3.4.3.18
pkg-resources==0.0.0
pyserial==3.4
python-engineio==2.3.2
python-socketio==2.0.0
six==1.11.0
Werkzeug==0.14.1
